@@ -1,77 +1,55 @@
 {
-  "name": "docker-manager-page",
-  "version": "4.4.0",
-  "description": "A Docker Manager with Element UI & axios & iconfont & permission control & lint",
-  "author": "Pan <panfree23@gmail.com>",
+  "name": "docker-manager-ui",
+  "version": "1.5.0",
   "scripts": {
-    "lint": "eslint --ext .js,.vue src",
-    "build:prod": "vue-cli-service build",
-    "build:stage": "vue-cli-service build --mode staging",
-    "dev": "vue-cli-service serve",
-    "i18n:report": "vue-cli-service i18n:report --src \"./src/**/*.?(js|vue)\" --locales \"./src/locales/**/*.json\"",
-    "preview": "node build/index.js --preview",
-    "svgo": "svgo -f src/icons/svg --config=src/icons/svgo.yml",
-    "test:ci": "npm run lint && npm run test:unit",
-    "test:unit": "jest --clearCache && vue-cli-service test:unit"
+    "dev": "vite",
+    "start": "vite",
+    "build": "vite build --mode=production",
+    "build:stag": "vite build --mode=staging",
+    "serve": "vite preview"
   },
   "dependencies": {
-    "axios": "0.18.1",
-    "codemirror": "^5.64.0",
-    "core-js": "3.6.5",
-    "element-ui": "2.13.2",
-    "js-cookie": "2.2.0",
-    "js-yaml": "^4.1.0",
-    "jsonlint": "^1.6.3",
-    "normalize.css": "^7.0.0",
+    "@element-plus/icons": "^0.0.11",
+    "@element-plus/icons-vue": "^2.0.10",
+    "@kangc/v-md-editor": "^2.3.5",
+    "@vueuse/core": "^8.0.0",
+    "axios": "^1.2.2",
+    "cropperjs": "^1.5.12",
+    "echarts": "^5.3.1",
+    "element-plus": "^2.2.28",
+    "jsencrypt": "^3.3.1",
+    "mockjs": "^1.1.0",
+    "monaco-editor": "^0.34.1",
+    "normalize.css": "^8.0.1",
     "nprogress": "^0.2.0",
-    "path-to-regexp": "2.4.0",
-    "script-loader": "^0.7.2",
-    "vue": "2.6.10",
+    "sortablejs": "^1.14.0",
+    "splitpanes": "^3.1.1",
+    "throttle-debounce": "^3.0.1",
+    "v-contextmenu": "^3.0.0",
+    "vite-plugin-monaco-editor": "^1.1.0",
+    "vue": "^3.2.31",
     "vue-count-to": "^1.0.13",
-    "vue-i18n": "^8.26.7",
-    "vue-i18n-bridge": "^9.2.0-beta.10",
-    "vue-router": "3.0.6",
-    "vuex": "3.1.0",
-    "xterm": "^4.16.0",
-    "xterm-addon-attach": "^0.6.0",
-    "xterm-addon-fit": "^0.5.0",
-    "xterm-addon-web-links": "^0.5.0"
+    "vue-i18n": "9.1.10",
+    "vue-router": "4",
+    "vue3-count-to": "^1.1.2",
+    "vuex": "^4.0.2",
+    "xlsx": "^0.18.3",
+    "xterm": "^5.1.0",
+    "xterm-addon-attach": "^0.8.0",
+    "xterm-addon-fit": "^0.7.0",
+    "xterm-addon-web-links": "^0.8.0"
   },
   "devDependencies": {
-    "@intlify/vue-i18n-loader": "^3.2.0",
-    "@vue/cli-plugin-babel": "4.4.4",
-    "@vue/cli-plugin-eslint": "4.4.4",
-    "@vue/cli-plugin-unit-jest": "4.4.4",
-    "@vue/cli-service": "4.4.4",
-    "@vue/test-utils": "1.0.0-beta.29",
-    "autoprefixer": "9.5.1",
-    "babel-eslint": "10.1.0",
-    "babel-jest": "23.6.0",
-    "babel-plugin-dynamic-import-node": "2.3.3",
-    "chalk": "2.4.2",
-    "connect": "3.6.6",
-    "eslint": "6.7.2",
-    "eslint-plugin-vue": "6.2.2",
-    "html-webpack-plugin": "3.2.0",
-    "jsencrypt": "^3.2.1",
-    "mockjs": "1.0.1-beta3",
-    "runjs": "4.3.2",
-    "sass": "1.26.8",
-    "sass-loader": "8.0.2",
-    "script-ext-html-webpack-plugin": "2.1.3",
-    "serve-static": "1.13.2",
-    "svg-sprite-loader": "4.1.3",
-    "svgo": "1.2.2",
-    "vue-cli-plugin-i18n": "^2.3.1",
-    "vue-template-compiler": "2.6.10"
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions"
-  ],
-  "engines": {
-    "node": ">=8.9",
-    "npm": ">= 3.0.0"
-  },
-  "license": "MIT"
+    "@types/node": "^17.0.21",
+    "@vitejs/plugin-vue": "^4.0.0",
+    "@vue/compiler-sfc": "^3.2.31",
+    "eslint": "^8.11.0",
+    "sass": "^1.49.9",
+    "typescript": "^4.6.2",
+    "unplugin-auto-import": "^0.6.4",
+    "unplugin-vue-components": "^0.18.0",
+    "vite": "^4.0.0",
+    "vite-plugin-mock": "2.9.6",
+    "vue-tsc": "^0.32.1"
+  }
 }
