@@ -1,96 +1,145 @@
 <template>
-  <div :class="classObj" class="app-wrapper">
-    <div v-if="device==='mobile'&&sidebar.opened" class="drawer-bg" @click="handleClickOutside" />
-    <sidebar class="sidebar-container" />
-    <div class="main-container">
-      <div :class="{'fixed-header':fixedHeader}">
-        <navbar />
-      </div>
-      <app-main />
-    </div>
-  </div>
+  <el-container style="height: 100vh">
+    <div
+      class="mask"
+      v-show="!isCollapse && !contentFullScreen"
+      @click="hideMenu"
+    ></div>
+    <el-aside
+      :width="isCollapse ? '60px' : '250px'"
+      :class="isCollapse ? 'hide-aside' : 'show-side'"
+      v-show="!contentFullScreen"
+    >
+      <Logo v-if="showLogo" />
+      <Menu />
+    </el-aside>
+    <el-container>
+      <el-header v-show="!contentFullScreen">
+        <Header />
+      </el-header>
+      <Tabs v-show="showTabs" />
+      <el-main>
+        <router-view v-slot="{ Component, route }">
+          <transition
+            :name="route.meta.transition || 'fade-transform'"
+            mode="out-in"
+          >
+            <keep-alive
+              v-if="keepAliveComponentsName"
+              :include="keepAliveComponentsName"
+            >
+              <component :is="Component" :key="route.fullPath" />
+            </keep-alive>
+            <component v-else :is="Component" :key="route.fullPath" />
+          </transition>
+        </router-view>
+      </el-main>
+    </el-container>
+  </el-container>
 </template>
 
-<script>
-import { Navbar, Sidebar, AppMain } from './components'
-import ResizeMixin from './mixin/ResizeHandler'
-
-export default {
-  name: 'Layout',
+<script lang="ts">
+import { defineComponent, computed, onBeforeMount } from "vue";
+import { useStore } from "vuex";
+import { useRouter } from "vue-router";
+import { useEventListener } from "@vueuse/core";
+import Menu from "./Menu/index.vue";
+import Logo from "./Logo/index.vue";
+import Header from "./Header/index.vue";
+import Tabs from "./Tabs/index.vue";
+export default defineComponent({
   components: {
-    Navbar,
-    Sidebar,
-    AppMain
+    Menu,
+    Logo,
+    Header,
+    Tabs,
   },
-  mixins: [ResizeMixin],
-  computed: {
-    sidebar() {
-      return this.$store.state.app.sidebar
-    },
-    device() {
-      return this.$store.state.app.device
-    },
-    fixedHeader() {
-      return this.$store.state.settings.fixedHeader
-    },
-    classObj() {
-      return {
-        hideSidebar: !this.sidebar.opened,
-        openSidebar: this.sidebar.opened,
-        withoutAnimation: this.sidebar.withoutAnimation,
-        mobile: this.device === 'mobile'
+  setup() {
+    const store = useStore();
+    // computed
+    const isCollapse = computed(() => store.state.app.isCollapse);
+    const contentFullScreen = computed(() => store.state.app.contentFullScreen);
+    const showLogo = computed(() => store.state.app.showLogo);
+    const showTabs = computed(() => store.state.app.showTabs);
+    const keepAliveComponentsName = computed(() => store.getters['keepAlive/keepAliveComponentsName']);
+    // 页面宽度变化监听后执行的方法
+    const resizeHandler = () => {
+      if (document.body.clientWidth <= 1000 && !isCollapse.value) {
+        store.commit("app/isCollapseChange", true);
+      } else if (document.body.clientWidth > 1000 && isCollapse.value) {
+        store.commit("app/isCollapseChange", false);
       }
-    }
+    };
+    // 初始化调用
+    resizeHandler();
+    // beforeMount
+    onBeforeMount(() => {
+      // 监听页面变化
+      useEventListener("resize", resizeHandler);
+    });
+    // methods
+    // 隐藏菜单
+    const hideMenu = () => {
+      store.commit("app/isCollapseChange", true);
+    };
+    return {
+      isCollapse,
+      showLogo,
+      showTabs,
+      contentFullScreen,
+      keepAliveComponentsName,
+      hideMenu,
+    };
   },
-  methods: {
-    handleClickOutside() {
-      this.$store.dispatch('app/closeSideBar', { withoutAnimation: false })
-    }
-  }
-}
+});
 </script>
 
 <style lang="scss" scoped>
-  @import "~@/styles/mixin.scss";
-  @import "~@/styles/variables.scss";
-
-  .main-container{
-    height: 100%;
+.el-header {
+  padding-left: 0;
+  padding-right: 0;
+}
+.el-aside {
+  display: flex;
+  flex-direction: column;
+  transition: 0.2s;
+  overflow-x: hidden;
+  transition: 0.3s;
+  &::-webkit-scrollbar {
+    width: 0 !important;
   }
-  .app-wrapper {
-    @include clearfix;
-    position: relative;
-    height: 100%;
-    width: 100%;
-    &.mobile.openSidebar{
-      position: fixed;
-      top: 0;
-    }
-  }
-  .drawer-bg {
-    background: #000;
-    opacity: 0.3;
-    width: 100%;
-    top: 0;
-    height: 100%;
-    position: absolute;
-    z-index: 999;
-  }
-
-  .fixed-header {
+}
+.el-main {
+  background-color: var(--system-container-background);
+  height: 100%;
+  padding: 0;
+  overflow-x: hidden;
+}
+:deep(.el-main-box) {
+  width: 100%;
+  height: 100%;
+  overflow-y: auto;
+  box-sizing: border-box;
+}
+@media screen and (max-width: 1000px) {
+  .el-aside {
     position: fixed;
     top: 0;
-    right: 0;
-    z-index: 9;
-    width: calc(100% - #{$sideBarWidth});
-    transition: width 0.28s;
+    left: 0;
+    height: 100vh;
+    z-index: 1000;
+    &.hide-aside {
+      left: -250px;
+    }
   }
-
-  .hideSidebar .fixed-header {
-    width: calc(100% - 54px)
+  .mask {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100vw;
+    height: 100vh;
+    z-index: 999;
+    background: rgba(0, 0, 0, 0.5);
   }
-
-  .mobile .fixed-header {
-    width: 100%;
-  }
+}
 </style>
