@@ -1,5 +1,5 @@
 import axios from 'axios'
-import { MessageBox, Message } from 'element-ui'
+import { ElMessageBox as MessageBox, ElMessage as Message } from 'element-plus'
 import store from '@/store'
 import { getToken } from '@/utils/auth'
 
@@ -15,7 +15,7 @@ const service = axios.create({
 
 // request interceptor
 service.interceptors.request.use(
-  config => {
+  (config) => {
     // do something before request is sent
 
     if (store.getters.token) {
@@ -26,7 +26,7 @@ service.interceptors.request.use(
     }
     return config
   },
-  error => {
+  (error) => {
     // do something with request error
     console.log('error:', error) // for debug
     return Promise.reject(error)
@@ -38,14 +38,14 @@ service.interceptors.response.use(
   /**
    * If you want to get http information such as headers or status
    * Please return  response => response
-  */
+   */
 
   /**
    * Determine the request status by custom code
    * Here is just an example
    * You can also judge the status by HTTP Status Code
    */
-  response => {
+  (response) => {
     const res = response.data
     // if the custom code is not 100200, it is judged as an error.
     if (res.code !== '100200') {
@@ -73,7 +73,7 @@ service.interceptors.response.use(
       return res
     }
   },
-  error => {
+  (error) => {
     Message({
       message: error + '',
       type: 'error',
